+++ resources/views/layouts/app.blade.php
@@ -34,6 +34,23 @@
                 <span>draab.at - <span title="{{config("app.deployDateTime")?->formatDateTimeWithSec()}}">{{config("app.version")}}</span></span>
             </footer>
         </div>
+        <script>
+            window.cookieStorage = {
+                getItem(key) {
+                    let cookies = document.cookie.split(";");
+                    for (let i = 0; i < cookies.length; i++) {
+                        let cookie = cookies[i].split("=");
+                        if (key == cookie[0].trim()) {
+                            return decodeURIComponent(cookie[1]);
+                        }
+                    }
+                    return null;
+                },
+                setItem(key, value) {
+                    document.cookie = key+' = '+encodeURIComponent(value)
+                }
+            }
+        </script>
         @livewireScripts
     </body>
 </html>